currently unable to seed reviews/user data. Commiting to have updated tables

+++ backend/db/models/review.js
@@ -27,7 +27,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     spotId: {
-      type:  DataTypes.INTEGER,
+      type: DataTypes.INTEGER,
       allowNull: false
     },
     review: {
@@ -35,7 +35,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     stars:{
-      type: DataTypes.INTEGER,
+      type: DataTypes.FLOAT,
     },
     imageId: {
       type: DataTypes.INTEGER
